Update InvestmentTransactionEditActivity.java
changed back to long

+++ app/src/main/java/com/money/manager/ex/investment/InvestmentTransactionEditActivity.java
@@ -316,7 +316,7 @@ public class InvestmentTransactionEditActivity
         // Handle action bar item clicks here. The action bar will
         // automatically e clicks on the Home/Up button, so long
         // as you specify a parent activity in AndroidManifest.xml.
-        int id = item.getItemId();
+        long id = item.getItemId();
 
         if (id == android.R.id.home) {
             return onActionCancelClick();
